requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.0.4
PyJWT==2.0.1
pymongo==3.11.2
python-decouple==3.4